Check root folder permissions. Skip existing directories and files. Add relevant tests.

+++ src/PackageGenerator.php
@@ -8,7 +8,12 @@ class PackageGenerator
         $validator = new ComplianceValidator();
         $lines = $validator->getFiles($root);
         $validatorResults = $validator->validate($lines);
-        $files = $this->createFiles($validatorResults, $root);
+        try {
+            $files = $this->createFiles($validatorResults, $root);
+        } catch (\Exception $e) {
+            echo $e->getMessage();
+        }
+
         $this->outputResults($files);
         return true;
     }
@@ -18,6 +23,10 @@ class PackageGenerator
         $root = $root ?: __DIR__ . '/../../../../';
         $root = realpath($root);
 
+        if (!is_writable($root) || !is_executable($root)) {
+            throw new GeneratorException("Cannot write into specified directory. Please check folder permissions for {$root}");
+        }
+
         $files = $this->createFileList($validatorResults);
         $createdFiles = [];
 
@@ -26,13 +35,15 @@ class PackageGenerator
             if ($isDir) {
                 $path = $root . '/' . substr($file, 0, -1);
                 $createdFiles[$file] = $path;
-                mkdir($path, 0755);
+                !is_dir($path) && !mkdir($path, 0755) && !is_dir($path);
                 continue;
             }
             $path = $root . '/' . $file . '.md';
             $createdFiles[$file] = $file . '.md';
-            file_put_contents($path, '');
-            chmod($path, 0644);
+            if (!file_exists($path)) {
+                file_put_contents($path, '');
+                chmod($path, 0644);
+            }
         }
 
         return $createdFiles;
@@ -58,4 +69,8 @@ class PackageGenerator
             echo "Created {$file}" . PHP_EOL;
         }
     }
+}
+
+class GeneratorException extends \Exception
+{
 }
